// Frontend/src/api/models/getReviewResponseDTO.ts
/**
 * Generated by orval v8.7.0 🍺
 * Do not edit manually.
 * AuditSupport
 * OpenAPI spec version: v1
 */

import type { EduProgramDTO } from "./eduProgramDTO";
import type { ProgramVersionDTO } from "./programVersionDTO";


export interface GetReviewResponseDTO {
  id: string;
  commentary?: string | null;
  isFinished: boolean;
  isSuccess: boolean;
  programVersion?: ProgramVersionDTO;
  program?: EduProgramDTO;
}